+++ public/src/components/channelManagement/bannerTests/bannerTestsForm.tsx
@@ -4,34 +4,8 @@ import { FrontendSettingsType } from '../../../utils/requests';
 import Sidebar from '../sidebar';
 import BannerTestEditor from './bannerTestEditor';
 import TestsFormLayout from '../testsFormLayout';
-import { ArticlesViewedSettings, Cta, Test, UserCohort } from '../helpers/shared';
-import { Region } from '../../../utils/models';
-
-export enum BannerTemplate {
-  ContributionsBanner = 'ContributionsBanner',
-  DigitalSubscriptionsBanner = 'DigitalSubscriptionsBanner',
-  GuardianWeeklyBanner = 'GuardianWeeklyBanner',
-}
-export interface BannerVariant {
-  name: string;
-  template: BannerTemplate;
-  heading?: string;
-  body: string;
-  highlightedText?: string;
-  cta?: Cta;
-  secondaryCta?: Cta;
-}
-
-export interface BannerTest extends Test {
-  name: string;
-  nickname?: string;
-  isOn: boolean;
-  minArticlesBeforeShowingBanner: number;
-  userCohort: UserCohort;
-  locations: Region[];
-  variants: BannerVariant[];
-  articlesViewedSettings?: ArticlesViewedSettings;
-}
+import { UserCohort } from '../helpers/shared';
+import { BannerTest } from '../../../models/banner';
 
 const createDefaultBannerTest = (newTestName: string, newTestNickname: string): BannerTest => ({
   name: newTestName,
